requirements.txt
numpy
tensorflow
pillow
fastapi
pydantic
typing
uvicorn